[uploader][s]: Add WordPress IDs and Names to authors.json.
NB: same authors.json will work on production because our staging is a
clone of production database so ids are all the same.

go through research/authors.json and use the data in
research/wordpress-team-authors.json to add `wordpress_id` and
`wordpress_name` wherever you can. So if you have an exact or close
match on id or name between value in authors.json and that in
wordpress-team-authors you can add the worpress stuff based on info in
wordpress-team-authors.json